backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
mysqlclient
cloudinary
python-dotenv
